+++ utils/http.js
@@ -3,7 +3,6 @@
 var cache = require('simples/lib/cache'),
 	domain = require('domain'),
 	path = require('path'),
-	stream = require('stream'),
 	url = require('url'),
 	utils = require('simples/utils/utils');
 
@@ -25,47 +24,55 @@ http.verbs = {
 	'PUT': 'put'
 };
 
-// Apply the selected route
 http.applyRoute = function (connection, route) {
-
-	var host = connection.parent;
-
-	// Check if any logger is defined and log data
-	if (host.logger.callback) {
-		utils.log(host, connection);
-	}
-
-	// Apply the route
 	if (typeof route === 'function') {
-		route.call(host, connection);
+		route.call(connection.parent, connection);
 	} else if (typeof route === 'string') {
 		connection.render(route);
-	} else {
+	}
+};
+
+// Apply the selected route
+http.prepareRoute = function (connection, route) {
+
+	var config = connection.parent.conf;
+
+	// Set the default keep alive timeout to 5 seconds
+	connection.keep(5000);
+
+	// Apply the route
+	if (utils.isObject(route)) {
 		http.applyStaticRoute(connection, route);
+	} else if (config.session.enabled) {
+		http.setSession(connection, function () {
+			http.applyRoute(connection, route);
+		});
+	} else {
+		http.applyRoute(connection, route);
 	}
 };
 
 // Apply the routes for static files and directories
-http.applyStaticRoute = function (connection, element) {
+http.applyStaticRoute = function (connection, route) {
 
 	var directory = null,
 		extension = '',
 		host = connection.parent,
 		mtime = 0,
 		routes = host.routes,
-		stats = element.stats;
+		stats = route.stats;
 
-	// Check if the element is a directory or a file
+	// Check if the route is a directory or a file
 	if (stats.isDirectory()) {
 
 		// Prepare directory items
-		directory = Object.keys(element.files).map(function (name) {
+		directory = Object.keys(route.files).map(function (name) {
 
 			var item = {};
 
 			// Prepare item object
 			item.name = name;
-			item.stats = element.files[name].stats;
+			item.stats = route.files[name].stats;
 
 			return item;
 		});
@@ -74,8 +81,8 @@ http.applyStaticRoute = function (connection, element) {
 		routes.serve.call(host, connection, directory);
 	} else {
 
-		// Prepare element properties
-		extension = path.extname(element.location).slice(1);
+		// Prepare route properties
+		extension = path.extname(route.location).slice(1);
 		mtime = stats.mtime.valueOf();
 
 		// Set the last modified time and the content type of the response
@@ -85,7 +92,7 @@ http.applyStaticRoute = function (connection, element) {
 		if (Number(connection.headers['if-modified-since']) === mtime) {
 			connection.status(304).end();
 		} else {
-			connection.end(element.content);
+			connection.end(route.content);
 		}
 	}
 };
@@ -93,18 +100,11 @@ http.applyStaticRoute = function (connection, element) {
 // Listener for HTTP requests
 http.connectionListener = function (host, request, response) {
 
-	var session = host.conf.session,
-		connection = new http.connection(host, request, response),
-		type = connection.type();
-
-	// Set the default keep alive timeout to 5 seconds
-	connection.keep(5000);
+	var connection = new http.connection(host, request, response);
 
 	// Prepare connection for routing
 	if (connection.method === 'OPTIONS') {
 		connection.end();
-	} else if (session.enabled && session.filter.test(type)) {
-		utils.getSession(host, connection, http.setSession);
 	} else {
 		http.routing(connection);
 	}
@@ -155,10 +155,14 @@ http.getDynamicRoute = function (connection, routes, verb) {
 // Check if the referer header is accepted by the host
 http.refers = function (host, connection) {
 
-	var headers = connection.headers,
-		hostname = url.parse(headers.referer || '').hostname || '',
+	var hostname = '',
 		referers = host.conf.referers,
 		valid = true;
+
+	// Check for the referer hostname
+	if (connection.headers.referer) {
+		hostname = url.parse(connection.headers.referer).hostname;
+	}
 
 	// Check if referer is found in the accepted referers list
 	if (referers.length && hostname !== connection.host) {
@@ -189,7 +193,7 @@ http.routing = function (connection) {
 			setImmediate(host.middlewares[index], connection, nextMiddleware);
 			index++;
 		} else if (!stop) {
-			http.applyRoute(connection, route);
+			http.prepareRoute(connection, route);
 		}
 	}
 
@@ -227,51 +231,47 @@ http.routing = function (connection) {
 	// Wrap the connection inside a domain
 	domain.create().on('error', function (error) {
 
-		// Emit the error to the host if it has any error listeners
-		if (host.listeners('error').length) {
-			host.emit('error', error);
-		} else {
-			console.error('\n' + error.stack + '\n');
-		}
+		// Emit safely the error to the host
+		utils.emitError(host, error, true);
 
-		// Try to apply the route for error 500
-		if (connection.status() !== 500) {
+		// Try to apply the route for error 500 or destroy the connection
+		if (!connection.started && connection.status() !== 500) {
 			connection.status(500);
 			this.bind(routes.error[500]).call(host, connection);
+		} else {
+			connection.destroy();
 		}
 	}).run(nextMiddleware);
 };
 
-// Set the session cookies for HTTP requests
-http.setSession = function (connection, session) {
+// Set the session cookies for HTTP connections
+http.setSession = function (connection, callback) {
 
-	var config = connection.parent.conf,
-		options = {},
-		store = config.session.store;
+	var host = connection.parent;
 
-	// Prepare cookies options
-	options.expires = config.session.timeout;
-	options.httpOnly = true;
+	// Prepare session object
+	utils.getSession(host, connection, function (connection, session) {
 
-	// Write the session cookies
-	connection.cookie('_session', session.id, options);
-	connection.cookie('_hash', session.hash, options);
+		var config = host.conf,
+			options = {};
 
-	// Link the session container to the connection
-	connection.session = session.container;
+		// Prepare cookies options
+		options.expires = config.session.timeout;
+		options.path = '/';
+		options.httpOnly = true;
 
-	// Write the session to the store and remove its reference
-	connection.on('finish', function () {
-		store.set(session.id, {
-			id: session.id,
-			hash: session.hash,
-			expire: config.session.timeout * 1000 + Date.now(),
-			container: connection.session
-		}, function () {
-			connection.session = null;
+		// Write the session cookies
+		connection.cookie('_session', session.id, options);
+		connection.cookie('_hash', session.hash, options);
+
+		// Link the session container to the connection
+		connection.session = session.container;
+
+		// Write the session to the store and remove its reference
+		connection.on('finish', function () {
+			utils.setSession(host, connection, session);
 		});
-	});
 
-	// Continue to process the request
-	http.routing(connection);
+		callback();
+	});
 };